Refactored code for Main
In this version, the Main class has been updated so that there aren't two objects, GoogleSheetsInitializer and GoogleSheetsService. The goal is to have GoogleSheetsInitializer be for testing the connection to the network and netcode operations, and GoogleSheetsService do basic CRUD (Create, Read, Update, Delete) operations.

All of this will be done in the next version through the use of an Interface class so that the website can make API calls.

Also, maybe make a third helper class called GoogleSheetsHelperFunctions to do more advanced functions if needed?

Current Goal: Refactor code and add interface. Then, make it so that every row in the GoogleSheet becomes an Object in an Array for easy scale and use.