Added interface for classes/Pt1 of commenting code
In this version, I made it so that everything in this program MUST interface with the GoogleSheetsApplicationInterface. Furthermore, I encapsulated .ServiceInitializer to make it responsible for all network connections. The .SheetsService is responsible for all data manipulation.

I haven't finished commenting out the functions since I need to go to class :(