Comments for tomorrow.
Next changes:

Choose specific row to modify and have those changes be put onto a stack.

When fetching all crops, user has to input which sheet it is.

Have a method to specificy where to start looking at the data? As in, if the parsing fails because the owners moved the data, have a call telling them which row and column the first piece of data is at!