Connected GoogleSheetsAPI to our program (#1)
* Added basic navigation

The navigation does not work, but the outline has been created. The case statements will be changed to the actual function later on!

* Changed minor things to prepare for Google Spreasheets

* We fully connected to the Google Sheets service!

This took wayyyy too long I'm happy it works. Now we need to learn how to edit data

* Changed minor wording so that connection makes more sense

* Added comments for increased readability

* Created GoogleSheetsService

This will be the main object that'll interact with the Google Sheet.

* Refactored code for Main

In this version, the Main class has been updated so that there aren't two objects, GoogleSheetsInitializer and GoogleSheetsService. The goal is to have GoogleSheetsInitializer be for testing the connection to the network and netcode operations, and GoogleSheetsService do basic CRUD (Create, Read, Update, Delete) operations.

All of this will be done in the next version through the use of an Interface class so that the website can make API calls.

Also, maybe make a third helper class called GoogleSheetsHelperFunctions to do more advanced functions if needed?

Current Goal: Refactor code and add interface. Then, make it so that every row in the GoogleSheet becomes an Object in an Array for easy scale and use.

* Added interface for classes/Pt1 of commenting code

In this version, I made it so that everything in this program MUST interface with the GoogleSheetsApplicationInterface. Furthermore, I encapsulated .ServiceInitializer to make it responsible for all network connections. The .SheetsService is responsible for all data manipulation.

I haven't finished commenting out the functions since I need to go to class :(

* Added boilerplate code for getItemsInSheet

Still need to do the same for the other methods!

* I can now get see all crops!

getItemsInSheet works now!

* Comments for tomorrow.

Next changes:

Choose specific row to modify and have those changes be put onto a stack.

When fetching all crops, user has to input which sheet it is.

Have a method to specificy where to start looking at the data? As in, if the parsing fails because the owners moved the data, have a call telling them which row and column the first piece of data is at!

* Can now add crops!

Basic functionality built, now need to make it so they can add specific data to the sheet.

* Can (almost) edit specific crops

It works!! The changes get staged to a queue, and only applied when they're pushed to the server.

I said almost because Right now we can only modify the cropName, and nothing else but will change in the next version!

* Many changes...

SheetName is no longer a param, it's built into every object. The goal is so that we can have rows from different sheets be pushed together.

Not sure if manage and pushChanges work yet, but will report soon!

* Note: Push changes does work!

Now I have to build out modifyData, build out ExceptionHandling, and should be done???

* Updated displayCrops method

I didn't like how displayCrops method was showing the crops as it looked like an array. Instead, I decided to make it more UI friendly by making the output look more like a table!

* Started building out modifyCrop...

PushChanges doesn't work, so I'm not sure if any of this actually works. Will report back...

* PushChanges method now works!

ONLY WORKS ON MODIFY CROP.

NOTHING HAPPENS ON ADD CROP (crop id not found, crop id updated).

NOTHING HAPPENS ON DELETE CROP.

WARNINGS:
Warning: Invalid numeric value for Crop ID: Crop ID
Warning: Invalid numeric value for Crop ID: Quantity Available

* deleteCrop kinda works???

It doesn't stage the changes, it just does it itself. Will fix in the next iteration.

* Added identifier, cropChanges, to Crop.

The goal of this is to tell the pushChanges method what happened to each crop, added, modified, or deleted. This is so that all changes can be staged and executed by one method in Main, and the appropriate method in GoogleSheetsService.

* PushChanges fully works!

Not sure what to say here but the program is technically now functionally complete. Now need to clean it up!